another big refactoring

+++ dvdtheque-business/src/main/java/fr/fredos/dvdtheque/service/dto/DvdDto.java
@@ -10,7 +10,7 @@ public class DvdDto implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	private java.lang.Integer id;
+	private Long id;
 	// fields
 	private Integer annee;
 	private Integer zone;
@@ -27,11 +27,11 @@ public class DvdDto implements Serializable {
 		this.zone = zone;
 	}
 
-	public java.lang.Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(java.lang.Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
