+++ dvdtheque-dao/src/main/java/fr/fredos/dvdtheque/dao/model/repository/FilmDao.java
@@ -8,12 +8,12 @@ import fr.fredos.dvdtheque.dao.model.object.Dvd;
 import fr.fredos.dvdtheque.dao.model.object.Film;
 
 public interface FilmDao {
-	public Film findFilm(Integer id);
+	public Film findFilm(Long id);
 	public Film findFilmByTitre(String titre);
-	public Film findFilmWithAllObjectGraph(Integer id);
-	public Integer saveNewFilm(Film film);
+	public Film findFilmWithAllObjectGraph(Long id);
+	public Long saveNewFilm(Film film);
 	public void updateFilm(Film film);
-	public Integer saveDvd(Dvd dvd);
+	public Long saveDvd(Dvd dvd);
 	public List<Film> findAllFilms();
 	public Set<Long> findAllTmdbFilms(Set<Long> tmdbIds);
 	public List<Film> findAllFilmsByCriteria(FilmFilterCriteriaDto filmFilterCriteriaDto);
