+++ dvdtheque-dao/src/main/java/fr/fredos/dvdtheque/dao/model/repository/PersonneDao.java
@@ -7,9 +7,9 @@ import fr.fredos.dvdtheque.dao.model.object.Personne;
 
 public interface PersonneDao {
 	
-	public Personne findByPersonneId(Integer id);
-	public Personne getPersonne(Integer id);
-	public Personne loadPersonne(Integer id);
+	public Personne findByPersonneId(Long id);
+	public Personne getPersonne(Long id);
+	public Personne loadPersonne(Long id);
 	public Personne findRealisateurByFilm(Film film);
 	public List<Personne> findAllRealisateur();
 	public List<Personne> findAllActeur();
@@ -17,7 +17,7 @@ public interface PersonneDao {
 	public List<Personne> findAllPersonneByType(Integer typeId);
 	public Personne findPersonneByName(String nom);
     public List<Personne> findAllPersonneByFilm(Film film);
-	public Integer savePersonne(Personne p);
+	public Long savePersonne(Personne p);
 	public void updatePersonne(Personne p);
 	public void deletePersonne(Personne p);
 	public void cleanAllPersons();
